Command failed: /Applications/Onlook.app/Contents/Resources/bun/bun-aarch64 run dev $ vite node:internal/process/promises:289             triggerUncaughtException(err, true /* fromPromise */);             ^
[Failed to load PostCSS config: Failed to load PostCSS config (searchPath: /Users/yifan/Documents/Onlook/Projects/project-1749040001955): [ReferenceError] module is not defined in ES module scope
This file is being treated as an ES module because it has a '.js' file extension and '/Users/yifan/Documents/Onlook/Projects/project-1749040001955/package.json' contains "type": "module". To treat it as a CommonJS script, rename it to use the '.cjs' file extension.
ReferenceError: module is not defined in ES module scope
This file is being treated as an ES module because it has a '.js' file extension and '/Users/yifan/Documents/Onlook/Projects/project-1749040001955/package.json' contains "type": "module". To treat it as a CommonJS script, rename it to use the '.cjs' file extension.
    at file:///Users/yifan/Documents/Onlook/Projects/project-1749040001955/postcss.config.js:1:1
    at ModuleJob.run (node:internal/modules/esm/module_job:222:25)
    at async ModuleLoader.import (node:internal/modules/esm/loader:316:24)
    at async importDefault (file:///Users/yifan/Documents/Onlook/Projects/project-1749040001955/node_modules/vite/dist/node/chunks/dep-C6uTJdX2.js:33759:18)
    at async Object.search (file:///Users/yifan/Documents/Onlook/Projects/project-1749040001955/node_modules/vite/dist/node/chunks/dep-C6uTJdX2.js:25915:23)]

Node.js v20.13.1
error: script "dev" exited with code 1